# 修正woc openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "修正woc",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/academy/captain": {
      "post": {
        "summary": "创建队长账号(新)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "comId": {
                    "type": "string"
                  }
                },
                "x-apifox-orders": [
                  "userCode",
                  "comId"
                ],
                "required": [
                  "userCode",
                  "comId"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "developing",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263915828-run"
      }
    },
    "/academy/team": {
      "get": {
        "summary": "获取队伍列表（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [
          {
            "name": "comId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamModel"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "errMsg",
                    "errCode"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845403-run"
      }
    },
    "/academy/judge": {
      "post": {
        "summary": "创建评委",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "comId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode",
                  "comId"
                ],
                "x-apifox-orders": [
                  "userCode",
                  "comId"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845404-run"
      },
      "delete": {
        "summary": "删除评委",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode"
                ],
                "x-apifox-orders": [
                  "userCode"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845405-run"
      },
      "get": {
        "summary": "评委列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [
          {
            "name": "comId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845406-run"
      }
    },
    "/judge": {
      "post": {
        "summary": "评分",
        "deprecated": false,
        "description": "",
        "tags": [
          "judge"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Score"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/judge",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845407-run"
      },
      "patch": {
        "summary": "修改评分",
        "deprecated": false,
        "description": "",
        "tags": [
          "judge"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Score"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/judge",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845408-run"
      }
    },
...
    "/captain/team": {
      "post": {
        "summary": "更新队伍信息（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "captain"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TeamModel"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "errMsg",
                    "errCode"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/captain",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845415-run"
      },
      "get": {
        "summary": "获取队伍信息（必做）",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TeamModel"
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845399-run"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "comId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "captainId": {
            "type": "integer"
          },
          "captainName": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "instructorIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "createTime": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "memberIds",
          "instructorIds",
          "status",
          "createTime"
        ],
        "x-apifox-orders": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "status",
          "memberIds",
          "instructorIds",
          "createTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "TeamModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "comId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "captainId": {
            "type": "integer"
          },
          "captainName": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "memberNames": {
            "type": "string",
            "description": "学号加姓名格式\n例如：B24XXXXXX某某某"
          },
          "instructorNames": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "memberNames",
          "instructorNames",
          "status"
        ],
        "x-apifox-orders": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "status",
          "memberNames",
          "instructorNames"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "GlobalResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errCode": {
            "type": "integer"
          },
          "errMsg": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {},
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "success",
          "data",
          "errMsg",
          "errCode"
        ],
        "x-apifox-orders": [
          "success",
          "errCode",
          "errMsg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "Score": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "judgeId": {
            "type": "string"
          },
          "workCode": {
            "type": "string"
          },
          "teamId": {
            "type": "integer"
          },
          "comId": {
            "type": "integer"
          },
          "score": {
            "type": "integer"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "judgeId",
          "workCode",
          "teamId",
          "comId",
          "score",
          "comment"
        ],
        "x-apifox-orders": [
          "id",
          "judgeId",
          "workCode",
          "teamId",
          "comId",
          "score",
          "comment"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}